feat: add authorization business rules to account management

+++ src/api/WalletUp.Application/Account/Commands/DeleteAccount/DeleteAccountCommandHandler.cs
@@ -1,18 +1,27 @@
 using WalletUp.Domain.Common;
 using WalletUp.Domain.Repositories;
 using MediatR;
+using WalletUp.Application.Common.Services;
 
 namespace WalletUp.Application.Account.Commands.DeleteAccount;
 
 public class DeleteAccountCommandHandler(
-    IAccountRepository accountRepository)
+    IAccountRepository accountRepository,
+    IUserContext userContext)
     :IRequestHandler<DeleteAccountCommand,Result>
 {
     public async Task<Result> Handle(DeleteAccountCommand request, CancellationToken cancellationToken)
     {
-         accountRepository.Delete(request.AccountId);
-         var affectedRows = await  accountRepository.SaveChanges();
-         
-        return affectedRows > 0 ? Result.Success() : Result.Failure(Errors.AccountNotFound);
+        var userId= userContext.UserId;
+        var account = accountRepository.GetAccountById(request.AccountId);
+        var canDelete=account.CanDelete(userId);
+        if (canDelete)
+        {
+            accountRepository.Delete(request.AccountId);
+            return Result.Success();
+        }
+        
+        return Errors.Forbidden;
+        
     }
 }
